Reclassify test as unit
