+++ kevm-pyk/src/tests/unit/test_unit.py
@@ -153,3 +153,69 @@ def test_method_sig_from_abi() -> None:
     assert method_sig_from_abi(json.loads(test2_abi)) == 'test2()'
     assert method_sig_from_abi(json.loads(test3_abi)) == 'test3(uint32,uint256)'
     assert method_sig_from_abi(json.loads(test4_abi)) == 'test4((address,(address,bytes)[3])[5])'
+
+contract_json = """
+{
+  "abi": [
+    {
+      "inputs": [
+        {
+          "components": [
+            {
+              "internalType": "address",
+              "name": "target",
+              "type": "address"
+            },
+            {
+              "internalType": "bytes",
+              "name": "callData",
+              "type": "bytes"
+            }
+          ],
+          "internalType": "struct IMulticall3.Call[]",
+          "name": "calls",
+          "type": "tuple[]"
+        }
+      ],
+      "name": "aggregate",
+      "outputs": [
+        {
+          "internalType": "uint256",
+          "name": "blockNumber",
+          "type": "uint256"
+        },
+        {
+          "internalType": "bytes[]",
+          "name": "returnData",
+          "type": "bytes[]"
+        }
+      ],
+      "stateMutability": "payable",
+      "type": "function"
+    }
+  ],
+  "deployedBytecode": {
+    "object": "0x",
+    "sourceMap": "",
+    "linkReferences": {}
+  },
+  "methodIdentifiers": {
+    "aggregate((address,bytes)[])": "252dba42"
+  },
+  "ast": {
+    "absolutePath": "lib/forge-std/src/interfaces/IMulticall3.sol"
+  },
+  "id": 17
+}
+"""
+
+
+def test_contract_creation() -> None:
+    contract = Contract('TestContract', json.loads(contract_json), foundry=True)
+    assert len(contract.methods) == 1
+    method = contract.methods[0]
+    assert method.sort == KSort('TestContractMethod')
+    assert method.id == int('252dba42', 16)
+    assert method.arg_types == ('tuple[]',)
+    assert method.contract_name == 'TestContract'
+    assert method.payable
